chore: remove cache-like operation, rename swap proposal in proto, errors start with number 2
