+++ x/fswap/keeper/keeper_test.go
@@ -137,14 +137,6 @@ func (s *KeeperTestSuite) TestSwap() {
 			true,
 			sdkerrors.ErrInsufficientFunds,
 		},
-		"swap with the same from-denom and to-denom": {
-			s.accWithFromCoin,
-			sdk.NewCoin(s.swap.GetToDenom(), s.initBalance),
-			s.swap.GetToDenom(),
-			sdk.ZeroInt(),
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
 	}
 	for name, tc := range testCases {
 		s.Run(name, func() {
@@ -208,86 +200,6 @@ func (s *KeeperTestSuite) TestSwapAll() {
 			actualAmount := s.keeper.GetBalance(ctx, tc.from, s.swap.GetToDenom()).Amount
 			expectedAmount := tc.expectedBalanceWithoutMultiply.Mul(s.swap.SwapMultiple)
 			s.Require().Equal(expectedAmount, actualAmount)
-		})
-	}
-}
-
-func (s *KeeperTestSuite) TestSwapInit() {
-	testCases := map[string]struct {
-		req              types.Swap
-		shouldThrowError bool
-		expectedError    error
-	}{
-		"valid getSwapInit": {
-			types.Swap{
-				FromDenom:           "fromD",
-				ToDenom:             "toD",
-				AmountCapForToDenom: sdk.OneInt(),
-				SwapMultiple:        sdk.OneInt(),
-			},
-			false,
-			nil,
-		},
-		"invalid empty from-denom": {
-			types.Swap{
-				FromDenom:           "",
-				ToDenom:             "toD",
-				AmountCapForToDenom: sdk.OneInt(),
-				SwapMultiple:        sdk.OneInt(),
-			},
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
-		"invalid empty to-denom": {
-			types.Swap{
-				FromDenom:           "fromD",
-				ToDenom:             "",
-				AmountCapForToDenom: sdk.OneInt(),
-				SwapMultiple:        sdk.OneInt(),
-			},
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
-		"invalid zero amount cap for to-denom": {
-			types.Swap{
-				FromDenom:           "fromD",
-				ToDenom:             "toD",
-				AmountCapForToDenom: sdk.ZeroInt(),
-				SwapMultiple:        sdk.OneInt(),
-			},
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
-		"invalid zero swap-rate": {
-			types.Swap{
-				FromDenom:           "fromD",
-				ToDenom:             "toD",
-				AmountCapForToDenom: sdk.OneInt(),
-				SwapMultiple:        sdk.ZeroInt(),
-			},
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
-		"invalid the same from-denom and to-denom": {
-			types.Swap{
-				FromDenom:           "same",
-				ToDenom:             "same",
-				AmountCapForToDenom: sdk.OneInt(),
-				SwapMultiple:        sdk.OneInt(),
-			},
-			true,
-			sdkerrors.ErrInvalidRequest,
-		},
-	}
-	for name, tc := range testCases {
-		s.Run(name, func() {
-			ctx, _ := s.ctx.CacheContext()
-			err := s.keeper.MakeSwap(ctx, tc.req)
-			if tc.shouldThrowError {
-				s.Require().ErrorIs(err, tc.expectedError)
-				return
-			}
-			s.Require().NoError(err)
 		})
 	}
 }
