+++ x/fswap/keeper/proposal_test.go
@@ -1,0 +1,87 @@
+package keeper_test
+
+import (
+	sdk "github.com/Finschia/finschia-sdk/types"
+	sdkerrors "github.com/Finschia/finschia-sdk/types/errors"
+	"github.com/Finschia/finschia-sdk/x/fswap/types"
+)
+
+func (s *KeeperTestSuite) TestMakeSwapProposal() {
+	testCases := map[string]struct {
+		req              types.Swap
+		shouldThrowError bool
+		expectedError    error
+	}{
+		"valid getSwap": {
+			types.Swap{
+				FromDenom:           "fromD",
+				ToDenom:             "toD",
+				AmountCapForToDenom: sdk.OneInt(),
+				SwapMultiple:        sdk.OneInt(),
+			},
+			false,
+			nil,
+		},
+		"invalid empty from-denom": {
+			types.Swap{
+				FromDenom:           "",
+				ToDenom:             "toD",
+				AmountCapForToDenom: sdk.OneInt(),
+				SwapMultiple:        sdk.OneInt(),
+			},
+			true,
+			sdkerrors.ErrInvalidRequest,
+		},
+		"invalid empty to-denom": {
+			types.Swap{
+				FromDenom:           "fromD",
+				ToDenom:             "",
+				AmountCapForToDenom: sdk.OneInt(),
+				SwapMultiple:        sdk.OneInt(),
+			},
+			true,
+			sdkerrors.ErrInvalidRequest,
+		},
+		"invalid zero amount cap for to-denom": {
+			types.Swap{
+				FromDenom:           "fromD",
+				ToDenom:             "toD",
+				AmountCapForToDenom: sdk.ZeroInt(),
+				SwapMultiple:        sdk.OneInt(),
+			},
+			true,
+			sdkerrors.ErrInvalidRequest,
+		},
+		"invalid zero swap-rate": {
+			types.Swap{
+				FromDenom:           "fromD",
+				ToDenom:             "toD",
+				AmountCapForToDenom: sdk.OneInt(),
+				SwapMultiple:        sdk.ZeroInt(),
+			},
+			true,
+			sdkerrors.ErrInvalidRequest,
+		},
+		"invalid the same from-denom and to-denom": {
+			types.Swap{
+				FromDenom:           "same",
+				ToDenom:             "same",
+				AmountCapForToDenom: sdk.OneInt(),
+				SwapMultiple:        sdk.OneInt(),
+			},
+			true,
+			sdkerrors.ErrInvalidRequest,
+		},
+	}
+	for name, tc := range testCases {
+		s.Run(name, func() {
+			ctx, _ := s.ctx.CacheContext()
+			err := s.keeper.MakeSwap(ctx, tc.req)
+			if tc.shouldThrowError {
+				s.Require().ErrorIs(err, tc.expectedError)
+				return
+			}
+			s.Require().NoError(err)
+		})
+	}
+}
