+++ x/fswap/module.go
@@ -117,9 +117,7 @@ func (am AppModule) Name() string {
 }
 
 // Route returns the fswap module's message routing key.
-func (am AppModule) Route() sdk.Route {
-	return sdk.NewRoute(types.RouterKey, NewHandler(am.keeper))
-}
+func (AppModule) Route() sdk.Route { return sdk.Route{} }
 
 // QuerierRoute returns the fswap module's query routing key.
 func (AppModule) QuerierRoute() string { return types.QuerierRoute }
