+++ x/fswap/types/errors.go
@@ -8,8 +8,8 @@ import (
 
 // x/fswap module sentinel errors
 var (
-	ErrInvalidState                = sdkerrors.Register(ModuleName, 1100, "swap module invalid state")
-	ErrCanNotHaveMoreSwap          = sdkerrors.Register(ModuleName, 1101, "no more swap allowed")
-	ErrSwappedNotFound             = sdkerrors.Register(ModuleName, 1102, "swapped does not exist")
-	ErrExceedSwappableToCoinAmount = sdkerrors.Register(ModuleName, 1103, "exceed swappable to-coin amount")
+	ErrInvalidState                = sdkerrors.Register(ModuleName, 2, "swap module invalid state")
+	ErrCanNotHaveMoreSwap          = sdkerrors.Register(ModuleName, 3, "no more swap allowed")
+	ErrSwappedNotFound             = sdkerrors.Register(ModuleName, 4, "swapped does not exist")
+	ErrExceedSwappableToCoinAmount = sdkerrors.Register(ModuleName, 5, "exceed swappable to-coin amount")
 )
